DTAD/requirements.txt
torch==2.1.0
torchvision==0.16.0
numpy==1.26.3
pillow==10.2.0
transformers==4.31.0
git+https://github.com/openai/CLIP.git
huggingface-hub==0.25.0
timm==0.9.16
fairscale==0.4.13
argparse
scikit-learn
tqdm